+++ django_app/book/templatetags/comment.py
@@ -1,4 +1,3 @@
-import re
 from django import template
 
 register = template.Library()
@@ -12,13 +11,3 @@ def star(value):
 @register.filter
 def hide(value):
     return value.replace(value[4:], "*****")
-
-
-@register.filter
-def img_filter(value):
-    p = re.compile('ht{2}p')
-    rep = value.replace(value[:7], "")
-    m = p.search(rep)
-    if m:
-        return True
-    return False
