+++ django_app/templates/include/autocomplete.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load comment %}
+{% load profile %}
 
 <link rel="stylesheet" href="{% static 'css/book-register.css' %}">
 <ul class="autocomplete-box">
